requirements.txt
Flask==3.1.1
elevenlabs==2.6.0
tensorflow==2.19.0
keras==3.10.0
numpy>=1.24.0
pydub>=0.25.0
SpeechRecognition==3.14.3
scipy>=1.11.0
azure-ai-inference==1.0.0b1
azure-mgmt-storage==21.1.0